Leave ControlMode in comm namespace
As this is used by external applications such as the ROS driver,
I think it makes sense to keep the namespace for backwards compatibility.

It also seems a logical choice, as this is part of the actual protocol definition.
